requirements.txt
numpy==1.23.0
streamlit==1.10.0
tensorflow-cpu==2.9.1